+++ Codice Qt/FreedomElectra/macchinario.h
@@ -11,7 +11,7 @@ protected:
     bool compressore;
 
 public:
-    Macchinario(string, string, int, int, int, float, int, bool, Cliente* =0, SalaCompressori* =0, int =0);
+    Macchinario(string, string, int, float, float, float, int, bool, Cliente* =0, SalaCompressori* =0, int =0);
 
     // get
     virtual int getKw() const;
