+++ Codice Qt/FreedomElectra/salacompressori.cpp
@@ -1,6 +1,6 @@
 #include "salacompressori.h"
 
-SalaCompressori::SalaCompressori(string n, Cliente* cl, int k, int p, int press, int port, string i):nome(n), proprietario(cl), kwTot(k), portataTot(p), pressioneRichiesta(press), portataRichiesta(port), impianto(i){}
+SalaCompressori::SalaCompressori(string n, Cliente* cl, int k, int p, float press, int port, string i):nome(n), proprietario(cl), kwTot(k), portataTot(p), pressioneRichiesta(press), portataRichiesta(port), impianto(i){}
 
 SalaCompressori::~SalaCompressori(){}
 
@@ -23,7 +23,7 @@ vector<Componente*>& SalaCompressori::getComponenti() const {
 Componente* SalaCompressori::getComponente(int i) const{
     return const_cast<Componente*>(componenti[i]);
 }
-int SalaCompressori::getPressioneRichiesta() const {
+float SalaCompressori::getPressioneRichiesta() const {
     return pressioneRichiesta;
 }
 int SalaCompressori::getPortataRichiesta() const {
@@ -40,7 +40,7 @@ void SalaCompressori::setNome(string s){
 void SalaCompressori::setProprietario(Cliente * punt){
     proprietario = punt;
 }
-void SalaCompressori::setPressioneRichiesta(int p) {
+void SalaCompressori::setPressioneRichiesta(float p) {
     pressioneRichiesta = p;
 }
 void SalaCompressori::setPortataRichiesta(int p) {
